+++ src/tada-runners.ads
@@ -3,6 +3,8 @@ package Tada.Runners is
 
    function Run_GNATdoc (Project : String) return Boolean;
 
+   function Run_GNATformat (Project : String) return Boolean;
+
    function Run_GPRBuild (Project : String; Profile : String) return Boolean;
 
    function Run_Tar (Source : String; Target : String) return Boolean;